feat: More cleanup
