--- frontend/src/components/application/view/actions-menu.tsx
@@ -1,26 +1,0 @@
-import { MoreVertical } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import {
-	DropdownMenu,
-	DropdownMenuTrigger,
-	DropdownMenuContent,
-	DropdownMenuItem,
-} from "@/components/ui/dropdown-menu";
-
-export function ActionsMenu() {
-	return (
-		<DropdownMenu>
-			<DropdownMenuTrigger asChild onClick={(e) => e.preventDefault()}>
-				<Button variant="ghost" size="icon" className="h-8 w-8">
-					<MoreVertical className="h-4 w-4" />
-				</Button>
-			</DropdownMenuTrigger>
-			<DropdownMenuContent align="end">
-				<DropdownMenuItem>Sync</DropdownMenuItem>
-				<DropdownMenuItem>Refresh</DropdownMenuItem>
-				<DropdownMenuItem>Settings</DropdownMenuItem>
-				<DropdownMenuItem className="text-destructive">Delete</DropdownMenuItem>
-			</DropdownMenuContent>
-		</DropdownMenu>
-	);
-}
