+++ frontend/src/routes/_authenticated/applications/index.tsx
@@ -1,12 +1,27 @@
-import { ApplicationView } from "@/components/application/view/application-view";
 import UpsertApplicationDialog from "@/components/dialogs/upsert-application";
 import { HealthStatus, SyncStatus, type Application } from "@/lib/applications";
-import { createFileRoute } from "@tanstack/react-router";
+import { createFileRoute, Link } from "@tanstack/react-router";
 import { useState } from "react";
 import { Button } from "@/components/ui/button";
-import { LayoutGrid, List, RefreshCw, Search } from "lucide-react";
+import {
+	Box,
+	GitBranch,
+	GitCommit,
+	LayoutGrid,
+	List,
+	MoreVertical,
+	RefreshCw,
+	Search,
+} from "lucide-react";
 import { Input } from "@/components/ui/input";
 import { cn } from "@/lib/utils";
+import { StatusBadge } from "@/components/status-badge";
+import {
+	DropdownMenu,
+	DropdownMenuTrigger,
+	DropdownMenuContent,
+	DropdownMenuItem,
+} from "@/components/ui/dropdown-menu";
 
 export const Route = createFileRoute("/_authenticated/applications/")({
 	component: ApplicationsPage,
@@ -100,6 +115,24 @@ const mockApps: Application[] = [
 	},
 ];
 
+function ActionsMenu() {
+	return (
+		<DropdownMenu>
+			<DropdownMenuTrigger asChild onClick={(e) => e.preventDefault()}>
+				<Button variant="ghost" size="icon" className="h-8 w-8">
+					<MoreVertical className="h-4 w-4" />
+				</Button>
+			</DropdownMenuTrigger>
+			<DropdownMenuContent align="end">
+				<DropdownMenuItem>Sync</DropdownMenuItem>
+				<DropdownMenuItem>Refresh</DropdownMenuItem>
+				<DropdownMenuItem>Settings</DropdownMenuItem>
+				<DropdownMenuItem className="text-destructive">Delete</DropdownMenuItem>
+			</DropdownMenuContent>
+		</DropdownMenu>
+	);
+}
+
 function ApplicationsPage() {
 	const [viewMode, setViewMode] = useState<"grid" | "list">("grid");
 	const [searchQuery, setSearchQuery] = useState("");
@@ -181,7 +214,107 @@ function ApplicationsPage() {
 				</div>
 			</div>
 			<div>
-				<ApplicationView viewMode={viewMode} apps={filteredApps} />
+				{/* <ApplicationView viewMode={viewMode} apps={filteredApps} /> */}
+
+				{viewMode === "grid" ? (
+					<div className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 gap-4">
+						{filteredApps.map((app) => (
+							<Link
+								key={app.id}
+								to="/applications/$id"
+								params={{ id: app.id }}
+								className="group bg-card border border-border rounded-lg p-4 hover:border-primary transition-colors"
+							>
+								<div className="flex items-start justify-between">
+									<div className="flex items-center gap-3">
+										<div className="h-10 w-10 rounded-lg bg-primary/10 flex items-center justify-center">
+											<Box className="h-5 w-5 text-primary" />
+										</div>
+										<div>
+											<h3 className="font-medium group-hover:text-primary transition-colors">
+												{app.name}
+											</h3>
+										</div>
+									</div>
+									<ActionsMenu />
+								</div>
+
+								<div className="flex gap-2 mt-4">
+									<StatusBadge status={app.syncStatus} />
+									<StatusBadge status={app.healthStatus} />
+								</div>
+
+								<div className="mt-4 pt-4 border-t border-border space-y-2">
+									<div className="flex items-center gap-2 text-sm text-muted-foreground">
+										<GitBranch className="h-4 w-4" />
+										<span className="truncate">{app.repo}</span>
+									</div>
+									<div className="flex items-center justify-between text-sm">
+										<div className="flex items-center gap-2 text-muted-foreground">
+											<GitCommit className="h-4 w-4" />
+											<span>{app.commit}</span>
+										</div>
+										<span className="text-muted-foreground">{app.lastSync}</span>
+									</div>
+								</div>
+							</Link>
+						))}
+					</div>
+				) : (
+					<div className="bg-card border border-border rounded-lg overflow-hidden">
+						<table className="w-full">
+							<thead>
+								<tr className="border-b border-border">
+									<th className="text-left p-4 text-sm font-medium text-muted-foreground">Name</th>
+									<th className="text-left p-4 text-sm font-medium text-muted-foreground hidden sm:table-cell">
+										Project
+									</th>
+									<th className="text-left p-4 text-sm font-medium text-muted-foreground">
+										Status
+									</th>
+									<th className="text-left p-4 text-sm font-medium text-muted-foreground hidden md:table-cell">
+										Repository
+									</th>
+									<th className="text-left p-4 text-sm font-medium text-muted-foreground hidden lg:table-cell">
+										Last Sync
+									</th>
+									<th className="p-4"></th>
+								</tr>
+							</thead>
+							<tbody>
+								{filteredApps.map((app) => (
+									<tr
+										key={app.id}
+										className="border-b border-border last:border-0 hover:bg-muted/50"
+									>
+										<td className="p-4">
+											<Link
+												to="/applications/$id"
+												params={{ id: app.id }}
+												className="font-medium hover:text-primary"
+											>
+												{app.name}
+											</Link>
+										</td>
+										<td className="p-4">
+											<div className="flex gap-2 flex-wrap">
+												<StatusBadge status={app.syncStatus} />
+												<StatusBadge status={app.healthStatus} />
+											</div>
+										</td>
+										<td className="p-4 text-muted-foreground hidden md:table-cell">{app.repo}</td>
+										<td className="p-4 text-muted-foreground hidden lg:table-cell">
+											{app.lastSync}
+										</td>
+										<td className="p-4">
+											<ActionsMenu />
+										</td>
+									</tr>
+								))}
+							</tbody>
+						</table>
+					</div>
+				)}
 			</div>
 		</div>
 	);
